finish up PR
