+++ packages/www/src/components/Layout/MaxWidth.js
@@ -1,12 +1,8 @@
 import { h } from 'preact'
 
-const MaxWidth = ({ className, children }) => {
+const MaxWidth = ({ children }) => {
   return (
-    <div
-      className={`w-full py-0 px-5 md:px-8 max-w-screen-lg mx-auto ${
-        className ? className : ''
-      }`}
-    >
+    <div className={`w-full py-0 px-5 md:px-8 max-w-screen-lg mx-auto`}>
       {children}
     </div>
   )
