--- packages/www/src/components/Logo/Wordmark.js
@@ -1,9 +1,0 @@
-import { h } from 'preact'
-
-const Wordmark = () => {
-  return (
-    <p className='font-bold group-hover:text-blue-400'>Party Corgi Network</p>
-  )
-}
-
-export default Wordmark
