--- packages/www/src/components/Logo/index.js
@@ -1,14 +1,0 @@
-import { h } from 'preact'
-import PartyCorgi from './PartyCorgi.js'
-import Wordmark from './Wordmark.js'
-
-const Logo = () => {
-  return (
-    <a href='/' className='group flex items-center font-bold'>
-      <PartyCorgi />
-      <Wordmark />
-    </a>
-  )
-}
-
-export default Logo
