+++ packages/www/src/content/assets.mdx
@@ -1,0 +1,9 @@
+## party-corgi.gif
+120x120
+
+![Party Corgi](/party-corgi.gif)
+
+## party-corgi-large.gif
+500x500
+
+![Party Corgi Large](/party-corgi-large.gif)
