+++ packages/www/src/page-wrapper.js
@@ -1,23 +1,59 @@
 import { h } from 'preact'
 import { MDXProvider } from '@mdx-js/preact'
 import { Helmet } from 'react-helmet'
-import Header from './components/Header/index.js'
-import MDXComponents from './lib/MDXComponents.js'
+import TopBar from './components/Header.js'
+import MDXComponents from './mdx-components.js'
 
 const PageWrapper = ({ children, meta }) => {
   return (
-    <MDXProvider components={MDXComponents}>
+    <div>
       <Helmet>
         <title>{`${
           meta?.title ? meta?.title + ' | ' : ''
         }Party Corgi Network`}</title>
+        <meta
+          name='description'
+          content={'An inclusive community of practice helping each other grow'}
+        />
+        <meta property='og:title' content='Party Corgi Network' />,
+        <meta name='og:url' content='https://www.partycorgi.com/' />,
+        <meta
+          property='og:image'
+          content={`https://www.partycorgi.com/ogImage.jpg`}
+        />
+        <meta property='og:image:width' content='1200' />
+        <meta property='og:image:height' content='630' />
+        <meta property='og:image:alt' content='Party Corgi Network' />
+        <meta
+          property='og:description'
+          content={'An inclusive community of practice helping each other grow'}
+        />
+        <meta property='og:type' content='website' />
+        <meta name='twitter:card' content='summary_large_image' />
+        <meta name='twitter:creator' content='@partycorginetwk' />
+        <meta name='twitter:title' content='Party Corgi Network' />
+        <meta
+          name='twitter:image'
+          content={`https://www.partycorgi.com/ogImage.jpg`}
+        />
+        <meta
+          name='twitter:description'
+          content={'An inclusive community of practice helping each other grow'}
+        />
+        <link
+          rel='icon'
+          type='image/gif'
+          href={'/images/party-corgi-still.gif'}
+        />
         <link rel='stylesheet' href='/styles.css' />
       </Helmet>
       <div className='flex flex-col min-h-screen w-full bg-gray-900 text-white'>
-        <Header />
-        <div className='flex-1 mt-[70px]'>{children}</div>
+        <TopBar />
+        <div className='flex-1 mt-[70px]'>
+          <MDXProvider components={MDXComponents}>{children}</MDXProvider>
+        </div>
       </div>
-    </MDXProvider>
+    </div>
   )
 }
